+++ internal/ceres/inner_product_computer.h
@@ -129,10 +129,10 @@ class CERES_EXPORT_INTERNAL InnerProductComputer {
 
   CompressedRowSparseMatrix* CreateResultMatrix(
       const CompressedRowSparseMatrix::StorageType storage_type,
-      int num_nonzeros);
+      int64_t num_nonzeros);
 
-  int ComputeNonzeros(const std::vector<ProductTerm>& product_terms,
-                      std::vector<int>* row_block_nnz);
+  int64_t ComputeNonzeros(const std::vector<ProductTerm>& product_terms,
+                          std::vector<int>* row_block_nnz);
 
   void ComputeOffsetsAndCreateResultMatrix(
       const CompressedRowSparseMatrix::StorageType storage_type,
